refactor component naming
