+++ components/Leaderboard/MintNft/Education/data.json
