+++ components/Leaderboard/MintNft/Education/index.jsx
