+++ components/Leaderboard/MintNft/styles.jsx
